fjernet /teams/
forbedrede også statcard koden en smule, men det burde ikke ændre noget visuelt
