+++ frontend/src/components/Collections/collections.module.css
@@ -2,7 +2,7 @@
   display: flex;
   flex-direction: column;
   width: 100%;
-  height: auto;
+  height: min-content;
   background: white;
   border: 1.5px solid var(--rule);
   position: relative;
